:white_check_mark: [#1661] Test static vars registry
